Preliminary refactor
