+++ app/src/main/contrib/actions/contribTouchstonesActionCreators.ts
@@ -1,20 +1,20 @@
 import {Dispatch} from "redux";
 
 import {ContribAppState} from "../reducers/contribAppReducers";
-import {TouchstonesService} from "../services/TouchstonesService";
-import {SetCurrentTouchstoneVersion, TouchstonesFetched, TouchstoneTypes} from "../actionTypes/TouchstonesTypes";
+import {TouchstonesService} from "../../shared/services/TouchstonesService";
+import {SetCurrentTouchstoneVersion, TouchstonesFetchedForGroup, TouchstoneTypes} from "../../shared/actionTypes/TouchstonesTypes";
 import {Touchstone} from "../../shared/models/Generated";
 import {flatMap} from "../../shared/ArrayHelpers";
 
-export const touchstonesActionCreators = {
+export const contribTouchstonesActionCreators = {
 
     getTouchstonesByGroupId(groupId: string) {
         return async (dispatch: Dispatch<ContribAppState>, getState: () => ContribAppState) => {
             const touchstones: Touchstone[] = await (new TouchstonesService(dispatch, getState)).getTouchstonesByGroupId(groupId);
             return dispatch({
-                type: TouchstoneTypes.TOUCHSTONES_FETCHED,
+                type: TouchstoneTypes.TOUCHSTONES_FETCHED_FOR_GROUP,
                 data: touchstones
-            } as TouchstonesFetched );
+            } as TouchstonesFetchedForGroup );
         }
     },
