+++ app/src/main/contrib/actions/pages/chooseActionPageActionCreators.ts
@@ -3,7 +3,7 @@ import { Dispatch } from "redux";
 import { modellingGroupsActionCreators } from "../modellingGroupsActionCreators";
 import {breadcrumbsActionCreators} from "../../../shared/actions/breadcrumbsActionsCreators";
 import {ChooseActionPageComponent, ChooseActionPageLocationProps} from "../../components/ChooseAction/ChooseActionPage";
-import {touchstonesActionCreators} from "../touchstonesActionCreators";
+import {contribTouchstonesActionCreators} from "../contribTouchstonesActionCreators";
 import {chooseGroupPageActionCreators} from "./chooseGroupPageActionCreators";
 import {ContribAppState} from "../../reducers/contribAppReducers";
 
@@ -20,7 +20,7 @@ export const chooseActionPageActionCreators = {
         return async (dispatch: Dispatch<ContribAppState>) => {
             await dispatch(chooseGroupPageActionCreators.loadData());
             dispatch(modellingGroupsActionCreators.setCurrentGroup(props.groupId));
-            await dispatch(touchstonesActionCreators.getTouchstonesByGroupId(props.groupId));
+            await dispatch(contribTouchstonesActionCreators.getTouchstonesByGroupId(props.groupId));
         }
     }
