+++ app/src/main/contrib/reducers/contribTouchstonesReducer.ts
@@ -1,5 +1,5 @@
-import {TouchstonesAction, TouchstoneTypes} from "../actionTypes/TouchstonesTypes";
 import {Touchstone, TouchstoneVersion} from "../../shared/models/Generated";
+import {TouchstonesAction, TouchstoneTypes} from "../../shared/actionTypes/TouchstonesTypes";
 
 export interface TouchstonesState {
     touchstones: Touchstone[];
@@ -11,9 +11,9 @@ export const touchstonesInitialState: TouchstonesState = {
     currentTouchstoneVersion: null
 };
 
-export const touchstonesReducer = (state = touchstonesInitialState, action: TouchstonesAction): TouchstonesState => {
+export const contribTouchstonesReducer = (state = touchstonesInitialState, action: TouchstonesAction): TouchstonesState => {
     switch (action.type) {
-        case TouchstoneTypes.TOUCHSTONES_FETCHED:
+        case TouchstoneTypes.TOUCHSTONES_FETCHED_FOR_GROUP:
             return {...state, touchstones: action.data ? action.data : [] };
         case TouchstoneTypes.SET_CURRENT_TOUCHSTONE_VERSION:
             return {...state, currentTouchstoneVersion: action.data };
