+++ app/src/main/shared/actionTypes/TouchstonesTypes.ts
@@ -1,12 +1,12 @@
 import {Touchstone, TouchstoneVersion} from "../../shared/models/Generated";
 
 export enum TouchstoneTypes {
-    TOUCHSTONES_FETCHED = "TOUCHSTONES_FETCHED",
+    TOUCHSTONES_FETCHED_FOR_GROUP = "TOUCHSTONES_FETCHED_FOR_GROUP",
     SET_CURRENT_TOUCHSTONE_VERSION = "SET_CURRENT_TOUCHSTONE_VERSION"
 }
 
-export interface TouchstonesFetched {
-    type: TouchstoneTypes.TOUCHSTONES_FETCHED;
+export interface TouchstonesFetchedForGroup {
+    type: TouchstoneTypes.TOUCHSTONES_FETCHED_FOR_GROUP;
     data: Touchstone[];
 }
 
@@ -16,5 +16,5 @@ export interface SetCurrentTouchstoneVersion {
 }
 
 export type TouchstonesAction =
-    | TouchstonesFetched
+    | TouchstonesFetchedForGroup
     | SetCurrentTouchstoneVersion
