+++ app/src/test/contrib/actions/pages/ChooseActionPageActionsTests.ts
@@ -8,8 +8,8 @@ import {BreadcrumbsTypes} from "../../../../main/shared/actionTypes/BreadrumbsTy
 import {breadcrumbsModule} from "../../../../main/shared/modules/breadcrumbs";
 import {mockBreadcrumbs, mockModellingGroup, mockTouchstoneVersion} from "../../../mocks/mockModels";
 import {chooseActionPageActionCreators} from "../../../../main/contrib/actions/pages/chooseActionPageActionCreators";
-import {TouchstonesService} from "../../../../main/contrib/services/TouchstonesService";
-import {TouchstoneTypes} from "../../../../main/contrib/actionTypes/TouchstonesTypes";
+import {TouchstonesService} from "../../../../main/shared/services/TouchstonesService";
+import {TouchstoneTypes} from "../../../../main/shared/actionTypes/TouchstonesTypes";
 
 describe("Choose Action Page actions tests", () => {
     const sandbox = new Sandbox();
@@ -46,7 +46,7 @@ describe("Choose Action Page actions tests", () => {
             const expectedPayload = [
                 { type: ModellingGroupTypes.USER_GROUPS_FETCHED, data: [testGroup] },
                 { type: ModellingGroupTypes.SET_CURRENT_USER_GROUP, data: testGroup },
-                { type: TouchstoneTypes.TOUCHSTONES_FETCHED, data: [testTouchstone]},
+                { type: TouchstoneTypes.TOUCHSTONES_FETCHED_FOR_GROUP, data: [testTouchstone]},
                 { type: BreadcrumbsTypes.BREADCRUMBS_RECEIVED, data: testBreadcrumbs }
             ];
             expect(actions).to.eql(expectedPayload);
