+++ app/src/test/contrib/actions/pages/DownloadCoveragePageActionsTests.ts
@@ -14,8 +14,8 @@ import {
     mockResponsibilitySet,
     mockTouchstone
 } from "../../../mocks/mockModels";
-import {TouchstonesService} from "../../../../main/contrib/services/TouchstonesService";
-import {TouchstoneTypes} from "../../../../main/contrib/actionTypes/TouchstonesTypes";
+import {TouchstonesService} from "../../../../main/shared/services/TouchstonesService";
+import {TouchstoneTypes} from "../../../../main/shared/actionTypes/TouchstonesTypes";
 import {DiseasesService} from "../../../../main/contrib/services/DiseasesService";
 import {ResponsibilitiesService} from "../../../../main/contrib/services/ResponsibilitiesService";
 import {DiseasesTypes} from "../../../../main/contrib/actionTypes/DiseasesTypes";
@@ -25,7 +25,7 @@ import {downloadCoveragePageActionCreators} from "../../../../main/contrib/actio
 import {CoverageService} from "../../../../main/contrib/services/CoverageService";
 import {ScenarioAndCoverageSets} from "../../../../main/shared/models/Generated";
 import {CoverageTypes} from "../../../../main/contrib/actionTypes/CoverageTypes";
-import {TouchstonesState} from "../../../../main/contrib/reducers/touchstonesReducer";
+import {TouchstonesState} from "../../../../main/contrib/reducers/contribTouchstonesReducer";
 
 describe("Download Coverage Page actions tests", () => {
     const sandbox = new Sandbox();
@@ -100,7 +100,7 @@ describe("Download Coverage Page actions tests", () => {
             const expectedPayload = [
                 {type: ModellingGroupTypes.USER_GROUPS_FETCHED, data: [testGroup]},
                 {type: ModellingGroupTypes.SET_CURRENT_USER_GROUP, data: testGroup},
-                {type: TouchstoneTypes.TOUCHSTONES_FETCHED, data: [testTouchstone]},
+                {type: TouchstoneTypes.TOUCHSTONES_FETCHED_FOR_GROUP, data: [testTouchstone]},
                 {type: DiseasesTypes.DISEASES_FETCHED, data: [testDisease]},
                 {type: TouchstoneTypes.SET_CURRENT_TOUCHSTONE_VERSION, data: testTouchstoneVersion},
                 {type: ResponsibilitiesTypes.SET_RESPONSIBILITIES, data: testExtResponsibilitySet},
