+++ app/src/test/contrib/actions/pages/DownloadDemographicPageActionsTests.ts
@@ -10,8 +10,8 @@ import {
     mockBreadcrumbs, mockDemographicDataset, mockDisease, mockModellingGroup, mockResponsibilitySet, mockTouchstone,
     mockTouchstoneVersion
 } from "../../../mocks/mockModels";
-import {TouchstonesService} from "../../../../main/contrib/services/TouchstonesService";
-import {TouchstoneTypes} from "../../../../main/contrib/actionTypes/TouchstonesTypes";
+import {TouchstonesService} from "../../../../main/shared/services/TouchstonesService";
+import {TouchstoneTypes} from "../../../../main/shared/actionTypes/TouchstonesTypes";
 import {DiseasesService} from "../../../../main/contrib/services/DiseasesService";
 import {ResponsibilitiesService} from "../../../../main/contrib/services/ResponsibilitiesService";
 import {DiseasesTypes} from "../../../../main/contrib/actionTypes/DiseasesTypes";
@@ -20,7 +20,7 @@ import {ExtendedResponsibilitySet} from "../../../../main/contrib/models/Respons
 import {downloadDemographicsPageActionCreators} from "../../../../main/contrib/actions/pages/downloadDemographicsPageActionCreators";
 import {DemographicService} from "../../../../main/contrib/services/DemographicService";
 import {DemographicTypes} from "../../../../main/contrib/actionTypes/DemographicTypes";
-import {TouchstonesState} from "../../../../main/contrib/reducers/touchstonesReducer";
+import {TouchstonesState} from "../../../../main/contrib/reducers/contribTouchstonesReducer";
 
 describe("Download Demographic Page actions tests", () => {
     const sandbox = new Sandbox();
@@ -77,7 +77,7 @@ describe("Download Demographic Page actions tests", () => {
             const expectedPayload = [
                 { type: ModellingGroupTypes.USER_GROUPS_FETCHED, data: [testGroup] },
                 { type: ModellingGroupTypes.SET_CURRENT_USER_GROUP, data: testGroup },
-                { type: TouchstoneTypes.TOUCHSTONES_FETCHED, data: [testTouchstone]},
+                { type: TouchstoneTypes.TOUCHSTONES_FETCHED_FOR_GROUP, data: [testTouchstone]},
                 { type: DiseasesTypes.DISEASES_FETCHED, data: [testDisease]},
                 { type: TouchstoneTypes.SET_CURRENT_TOUCHSTONE_VERSION, data: testTouchstoneVersion},
                 { type: ResponsibilitiesTypes.SET_RESPONSIBILITIES, data: testExtResponsibilitySet},
