+++ app/src/test/contrib/actions/pages/ModelRunParametersPageActionsTests.ts
@@ -14,8 +14,8 @@ import {
     mockResponsibilitySet,
     mockTouchstone
 } from "../../../mocks/mockModels";
-import {TouchstonesService} from "../../../../main/contrib/services/TouchstonesService";
-import {TouchstoneTypes} from "../../../../main/contrib/actionTypes/TouchstonesTypes";
+import {TouchstonesService} from "../../../../main/shared/services/TouchstonesService";
+import {TouchstoneTypes} from "../../../../main/shared/actionTypes/TouchstonesTypes";
 import {DiseasesService} from "../../../../main/contrib/services/DiseasesService";
 import {ResponsibilitiesService} from "../../../../main/contrib/services/ResponsibilitiesService";
 import {DiseasesTypes} from "../../../../main/contrib/actionTypes/DiseasesTypes";
@@ -24,7 +24,7 @@ import {ExtendedResponsibilitySet} from "../../../../main/contrib/models/Respons
 import {modelRunParametersPageActionCreators} from "../../../../main/contrib/actions/pages/modelRunParametersPageActionCreators";
 import {RunParametersService} from "../../../../main/contrib/services/RunParametersService";
 import {RunParametersTypes} from "../../../../main/contrib/actionTypes/RunParametersTypes";
-import {TouchstonesState} from "../../../../main/contrib/reducers/touchstonesReducer";
+import {TouchstonesState} from "../../../../main/contrib/reducers/contribTouchstonesReducer";
 
 describe("Model Run Parameters Page actions tests", () => {
     const sandbox = new Sandbox();
@@ -81,7 +81,7 @@ describe("Model Run Parameters Page actions tests", () => {
             const expectedPayload = [
                 {type: ModellingGroupTypes.USER_GROUPS_FETCHED, data: [testGroup]},
                 {type: ModellingGroupTypes.SET_CURRENT_USER_GROUP, data: testGroup},
-                {type: TouchstoneTypes.TOUCHSTONES_FETCHED, data: [testTouchstone]},
+                {type: TouchstoneTypes.TOUCHSTONES_FETCHED_FOR_GROUP, data: [testTouchstone]},
                 {type: DiseasesTypes.DISEASES_FETCHED, data: [testDisease]},
                 {type: TouchstoneTypes.SET_CURRENT_TOUCHSTONE_VERSION, data: testTouchstoneVersion},
                 {type: ResponsibilitiesTypes.SET_RESPONSIBILITIES, data: testExtResponsibilitySet},
