+++ app/src/test/contrib/reducers/ContribTouchstonesReducerTests.ts
@@ -1,35 +1,35 @@
 import { expect } from "chai";
 
-import {touchstonesInitialState, touchstonesReducer} from "../../../main/contrib/reducers/touchstonesReducer";
-import { TouchstoneTypes } from "../../../main/contrib/actionTypes/TouchstonesTypes";
+import {touchstonesInitialState, contribTouchstonesReducer} from "../../../main/contrib/reducers/contribTouchstonesReducer";
+import { TouchstoneTypes } from "../../../main/shared/actionTypes/TouchstonesTypes";
 import {mockTouchstone, mockTouchstoneVersion} from "../../mocks/mockModels";
 
-describe('Touchstones reducer tests', () => {
+describe('Contrib touchstones reducer tests', () => {
     it('sets fetched touchstones', () => {
         const testTouchstone = mockTouchstone();
-        expect(touchstonesReducer(undefined, {
-            type: TouchstoneTypes.TOUCHSTONES_FETCHED,
+        expect(contribTouchstonesReducer(undefined, {
+            type: TouchstoneTypes.TOUCHSTONES_FETCHED_FOR_GROUP,
             data: [testTouchstone]
         })).to.eql({...touchstonesInitialState, touchstones: [testTouchstone]});
     });
 
     it('sets empty fetched touchstones', () => {
-        expect(touchstonesReducer(undefined, {
-            type: TouchstoneTypes.TOUCHSTONES_FETCHED,
+        expect(contribTouchstonesReducer(undefined, {
+            type: TouchstoneTypes.TOUCHSTONES_FETCHED_FOR_GROUP,
             data: null
         })).to.eql(touchstonesInitialState);
     });
 
     it('sets current touchstone', () => {
         const testTouchstoneVersion = mockTouchstoneVersion();
-        expect(touchstonesReducer(undefined, {
+        expect(contribTouchstonesReducer(undefined, {
             type: TouchstoneTypes.SET_CURRENT_TOUCHSTONE_VERSION,
             data: testTouchstoneVersion
         })).to.eql({...touchstonesInitialState, currentTouchstoneVersion: testTouchstoneVersion});
     });
 
     it('sets current touchstone empty', () => {
-        expect(touchstonesReducer(undefined, {
+        expect(contribTouchstonesReducer(undefined, {
             type: TouchstoneTypes.SET_CURRENT_TOUCHSTONE_VERSION,
             data: null
         })).to.eql(touchstonesInitialState);
