+++ app/src/test/shared/services/TouchstonesServiceTests.ts
@@ -2,7 +2,7 @@ import { expect } from "chai";
 import { createMemoryHistory } from 'history';
 
 import {createContribStore} from "../../../main/contrib/createStore";
-import {TouchstonesService} from "../../../main/contrib/services/TouchstonesService";
+import {TouchstonesService} from "../../../main/shared/services/TouchstonesService";
 import {Sandbox} from "../../Sandbox";
 import {ContribAppState} from "../../../main/contrib/reducers/contribAppReducers";
